restructured some file paths and added optional qaic_config override in compile
