+++ QEfficient/diffusers/models/__init__.py
@@ -4,7 +4,3 @@
 # SPDX-License-Identifier: BSD-3-Clause
 #
 # ----------------------------------------------------------------------------
-
-from QEfficient.diffusers.models.blocking_configurator import build_transformer_blocking_config
-
-__all__ = ["build_transformer_blocking_config"]
